Move commands into BookmarksCore
