+++ core/Sources/BookmarksCore/Commands/BookmarkDestructiveCommands.swift
@@ -20,8 +20,6 @@
 
 import SwiftUI
 
-import BookmarksCore
-
 struct BookmarkDesctructiveCommands: View {
 
     @Environment(\.manager) var manager: BookmarksManager
